requirements.txt
# Core LLM + Transformers
transformers==4.40.1
datasets==2.19.1
sentence-transformers==2.6.1
huggingface-hub==0.23.1

# LangChain RAG + Embeddings + Vector DB
langchain==0.1.19
faiss-cpu==1.8.0

# PDF and DOCX parsing
pdfplumber==0.11.0
python-docx==1.1.0

# Web App
streamlit==1.35.0

# Utilities
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1

# Optional: Fine-tuning + GPU
torch==2.2.2
accelerate==0.30.1
bitsandbytes==0.43.1

# Optional: Web APIs
fastapi==0.110.2
uvicorn==0.30.1

# Authentication CLI (optional)
huggingface-cli
